feat:added user

+++ posts/views.py
@@ -1,6 +1,6 @@
 from urllib.parse import quote_plus
 
-from django.http import HttpResponse ,HttpResponseRedirect
+from django.http import HttpResponse ,HttpResponseRedirect,Http404
 from django.shortcuts import render,get_object_or_404,redirect
 from django.core.paginator import Paginator,EmptyPage, PageNotAnInteger
 from .models import Post
@@ -10,9 +10,11 @@ from .forms import PostForm
 # Create your views here.
 def post_create(request):
     form = PostForm(request.POST or None,request.FILES or None)
-
+    if not request.user.is_staff or request.user.is_superuser:
+        raise Http404
     if form.is_valid()==True:
         instance=form.save(commit=False)
+        instance.user = request.user
         instance.save()
         messages.success(request,"created successfully")
         return HttpResponseRedirect(instance.get_absolute_url())
@@ -53,9 +55,10 @@ def post_detail(request,id=None):
 
 
 def post_update(request,id=None):
+    if not request.user.is_staff or request.user.is_superuser:
+        raise Http404
     instance = get_object_or_404(Post, id=id)
     form = PostForm(request.POST or None, request.FILES or None, instance=instance)
-
     if form.is_valid() == True:
         instance = form.save(commit=False)
         instance.save()
@@ -71,6 +74,8 @@ def post_update(request,id=None):
 
 
 def post_delete(request,id=None):
+    if not request.user.is_staff or request.user.is_superuser:
+        raise Http404
     instance = get_object_or_404(Post, id=id)
     instance.delete()
     return  redirect("posts:list")
